oe-layersetup: add a scarthgap chromium config
Add a chromium config for Scarthgap, just like we did for Kirkstone,
it is locked to meta-browser at Chromium v123 as we are not quite at
a point where we can float with head as we are carrying patches in
meta-arago to make things work,  however we are getting really close
to dropping this and floating with each Chromium release.

For Chromium to function correctly we need:
[meta-ti][scarthgap/master] ti-img-rogue-umlibs: bump SRCREV for alloc patch

from Randolph Sapp to have been picked up in meta-ti as this fixes an
allocation problem that Chromium exposes.

Also updating the commented out meta-browser lines to this same commit
in the mainstream scarthgap configs.

Signed-off-by: Darren Etheridge <detheridge@ti.com>
Acked-by: Chirag Shilwant <c-shilwant@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>